Simulation of Rician Fading Channel
This simulation demonstrates that the channel created is Rician Fading, which means that the antenna and end-user have at least one line of sight. The antenna propagates the signal in a scattering manner, and the signal reaches the user by reflecting off different surfaces, such as buildings, trees, vehicles, etc. If the channel is rician fading then at least one of the beam signals reaches the user without reflecting strengthening the signal to the user.